Adopt Vite+ (vp) for the frontend toolchain
- vite-plus 0.3, Vite 8 and laravel-vite-plugin 3.2; build/dev now run
  through vp, plugins wrapped in lazyPlugins() (matches the official
  Livewire starter kit).
- vp check (Oxfmt + Oxlint) scoped to the frontend sources; added
  check/check-fix scripts.
- Fix config()->float() on assistant.deadline (it was an int, which the
  typed getter rejects).

+++ config/assistant.php
@@ -31,7 +31,7 @@ return [
 
     // Hard cap across every attempt, so a chain of slow models can never reach
     // the gateway's timeout (which surfaces as a 504 to the visitor).
-    'deadline' => (int) env('ASSISTANT_DEADLINE', 45),
+    'deadline' => (float) env('ASSISTANT_DEADLINE', 45),
 
     // How many retrieved pages to pass as context, and how much of each.
     'sources' => (int) env('ASSISTANT_SOURCES', 8),
